+++ src/main/java/duke/command/AddTodoCommand.java
@@ -1,3 +1,12 @@
+package duke.command;
+
+import duke.errors.DukeException;
+import duke.main.Storage;
+import duke.main.Ui;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.task.ToDos;
+
 public class AddTodoCommand extends Command {
     private String description;
