+++ src/main/java/duke/command/ExitCommand.java
@@ -1,6 +1,15 @@
+package duke.command;
+
+import duke.errors.DukeException;
+import duke.main.Storage;
+import duke.main.Ui;
+import duke.task.TaskList;
+
 public class ExitCommand extends Command {
+
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
+
     }
 
     @Override
