+++ src/main/java/duke/errors/DukeException.java
@@ -1,3 +1,5 @@
+package duke.errors;
+
 public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
