+++ src/main/java/duke/task/ToDos.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 public class ToDos extends Task {
 
     public ToDos(String description) {
